refactor: extract shared WebAudio celebration/feedback utils
playTone and playCelebration were duplicated verbatim in Answer and
HangmanGame (~120 lines total). Both now come from utils/audio with
a single typed AudioContext accessor.
